server/pyproject.toml
[tool.poetry]
name = "blogsley"
version = "0.1.0"
description = ""
authors = ["Kurtis Fields <kurtisfields@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
ariadne = "^0.16.0"
tortoise-orm = {extras = ["aiosqlite", "asyncpg"], version = "^0.19.2"}
PyJWT = "^2.4.0"
python-slugify = "^6.1.2"
psycopg2 = "^2.8.6"
uvicorn = {extras = ["standard"], version = "^0.18.2"}
loguru = "^0.6.0"
pydantic = "^1.9.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
blogsley = 'blogsley.cli:entry_point'
